remove types.go files and place interfaces close to consumers
