+++ internal/xpkg/dep/resolver/image/resolver.go
@@ -47,6 +47,13 @@ type Resolver struct {
 	f Fetcher
 }
 
+// Fetcher defines how we expect to intract with the Image repository.
+type Fetcher interface {
+	Fetch(ctx context.Context, ref name.Reference, secrets ...string) (v1.Image, error)
+	Head(ctx context.Context, ref name.Reference, secrets ...string) (*v1.Descriptor, error)
+	Tags(ctx context.Context, ref name.Reference, secrets ...string) ([]string, error)
+}
+
 // NewResolver returns a new Resolver.
 func NewResolver(opts ...ResolverOption) *Resolver {
 	r := &Resolver{
