--- internal/xpkg/dep/resolver/image/types.go
@@ -1,29 +1,0 @@
-// Copyright 2021 Upbound Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package image
-
-import (
-	"context"
-
-	"github.com/google/go-containerregistry/pkg/name"
-	v1 "github.com/google/go-containerregistry/pkg/v1"
-)
-
-// Fetcher defines how we expect to intract with the Image repository.
-type Fetcher interface {
-	Fetch(ctx context.Context, ref name.Reference, secrets ...string) (v1.Image, error)
-	Head(ctx context.Context, ref name.Reference, secrets ...string) (*v1.Descriptor, error)
-	Tags(ctx context.Context, ref name.Reference, secrets ...string) ([]string, error)
-}
